models: consolidate durable accounting recovery in core
