+++ packages/agent-worker/tests/worker.test.ts
@@ -47,6 +47,7 @@ import {
 } from "@sixb/core/internal/agents"
 import { attachSixbErrorReporter } from "@sixb/core/internal/error-reporting"
 import { createSixbError } from "@sixb/core/internal/errors"
+import { enqueueAiModelCallRecovery } from "@sixb/core/internal/model-execution"
 import { bindRequestExecution } from "@sixb/core/internal/request-execution"
 import { workflowAgentStepActorId } from "@sixb/core/internal/workflows"
 import type {
@@ -83,7 +84,6 @@ import { prepareAgentAttachments } from "../src/attachments"
 import { AgentExecutionLostError, AgentFinalizationError } from "../src/errors"
 import { resolveAgentExecutionPlan } from "../src/execution-plan"
 import { finishRunOrThrow } from "../src/finalize"
-import { enqueueAiModelCallRecovery } from "../src/model-call-recovery"
 import { runAgentTurn } from "../src/run-agent-turn"
 import * as agentEnvironment from "../src/run-environment"
 import { createConversationAgentEnvironment } from "../src/run-environment"
